Add VoterInfoForm to VoterRegLookup
